web/requirements.txt
amqp==2.4.1
backcall==0.1.0
billiard==3.5.0.5
celery==4.2.1
Click==7.0
cycler==0.10.0
decorator==4.3.2
Django==2.1.5
django-filter==2.1.0
django-rq==2.0
djangorestframework==3.9.1
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.2
jupyter-client==5.2.4
jupyter-core==4.4.0
kiwisolver==1.0.1
kombu==4.3.0
Markdown==3.0.1
matplotlib==3.0.2
matplotlib-venn==0.11.5
mysqlclient==1.4.2
numpy==1.16.1
pandas==0.24.1
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.8
ptyprocess==0.6.0
Pygments==2.3.1
PyMySQL==0.9.3
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
pyzmq==17.1.2
redis==3.2.1
rq==1.0
scipy==1.2.0
seaborn==0.9.0
setuptools-scm==3.2.0
six==1.12.0
tornado==5.1.1
traitlets==4.3.2
vine==1.2.0
wcwidth==0.1.7
weblogo==3.7.1
